Refactor agent chat title flow
