+++ internal/agentapp/service.go
@@ -4,13 +4,11 @@ import (
 	"bytes"
 	"context"
 	"crypto/rand"
-	"database/sql"
 	"encoding/hex"
 	"encoding/json"
 	"errors"
 	"fmt"
 	"net/http"
-	"regexp"
 	"sort"
 	"strings"
 	"sync"
@@ -29,13 +27,7 @@ var (
 const (
 	maxToolArgumentsPreviewBytes = 2000
 	maxToolResultPreviewBytes    = 4000
-	titleReserveOutputTokens     = 64
-	maxGeneratedTitleRunes       = 50
 )
-
-const modelRequestPurposeTitle agent.ModelRequestPurpose = "title_generation"
-
-var thinkBlockPattern = regexp.MustCompile(`(?is)<think>.*?</think>\s*`)
 
 func IsBusy(err error) bool {
 	return errors.Is(err, ErrBusy)
@@ -84,18 +76,6 @@ func (s *Service) CreateConversation(ctx context.Context, scope Scope, title str
 
 func (s *Service) ListConversations(ctx context.Context, scope Scope) ([]platformdb.AgentConversation, error) {
 	return s.store.ListAgentConversations(ctx, scope.WorkspaceID, scope.PrincipalID)
-}
-
-func (s *Service) ConversationNeedsGeneratedTitle(ctx context.Context, scope Scope, conversationID string) (bool, error) {
-	conversation, err := s.store.GetAgentConversation(ctx, scope.WorkspaceID, scope.PrincipalID, conversationID)
-	if err != nil {
-		return false, err
-	}
-	if !isDefaultConversationTitle(conversation.Title) {
-		return false, nil
-	}
-	_, ok, err := firstUserPromptForTitle(ctx, s.store, scope, conversationID)
-	return ok, err
 }
 
 func (s *Service) ListMessages(ctx context.Context, scope Scope, conversationID string) ([]platformdb.AgentMessage, error) {
@@ -149,120 +129,6 @@ func (s *Service) ConversationTranscript(ctx context.Context, scope Scope, conve
 		return nil, err
 	}
 	return transcriptFromMessages(conversationID, messages), nil
-}
-
-func (s *Service) GenerateConversationTitle(ctx context.Context, scope Scope, conversationID string) (platformdb.AgentConversation, error) {
-	if !s.Enabled() {
-		return platformdb.AgentConversation{}, ErrDisabled
-	}
-	if s.store == nil {
-		return platformdb.AgentConversation{}, fmt.Errorf("agent store is required")
-	}
-	conversation, err := s.store.GetAgentConversation(ctx, scope.WorkspaceID, scope.PrincipalID, conversationID)
-	if err != nil {
-		return platformdb.AgentConversation{}, err
-	}
-	if !isDefaultConversationTitle(conversation.Title) {
-		return conversation, nil
-	}
-	firstPrompt, ok, err := firstUserPromptForTitle(ctx, s.store, scope, conversationID)
-	if err != nil {
-		return conversation, err
-	}
-	if !ok {
-		return conversation, nil
-	}
-	resp, err := s.model.Complete(ctx, agent.ModelRequest{
-		Purpose: modelRequestPurposeTitle,
-		Messages: []agent.Message{
-			{Role: agent.RoleSystem, Content: titleSystemPrompt()},
-			{Role: agent.RoleUser, Content: "Generate a title for this conversation:\n" + firstPrompt},
-		},
-		Tools:  nil,
-		Limits: agent.Limits{ReserveOutputTokens: titleReserveOutputTokens},
-	}, nil)
-	title := fallbackConversationTitle(firstPrompt)
-	if err == nil {
-		if generated := cleanGeneratedTitle(resp.Content); generated != "" {
-			title = generated
-		}
-	}
-	if title == "" {
-		return conversation, nil
-	}
-	latest, err := s.store.GetAgentConversation(ctx, scope.WorkspaceID, scope.PrincipalID, conversationID)
-	if err != nil {
-		return conversation, err
-	}
-	if !isDefaultConversationTitle(latest.Title) {
-		return latest, nil
-	}
-	updated, err := s.store.UpdateDefaultAgentConversationTitle(ctx, scope.WorkspaceID, scope.PrincipalID, conversationID, title)
-	if errors.Is(err, sql.ErrNoRows) {
-		return latest, nil
-	}
-	if err != nil {
-		return latest, err
-	}
-	return updated, nil
-}
-
-func firstUserPromptForTitle(ctx context.Context, store *platform.Store, scope Scope, conversationID string) (string, bool, error) {
-	messages, err := store.ListAgentMessages(ctx, scope.WorkspaceID, scope.PrincipalID, conversationID)
-	if err != nil {
-		return "", false, err
-	}
-	userCount := 0
-	firstPrompt := ""
-	for _, message := range messages {
-		if message.Role != platform.AgentMessageRoleUser {
-			continue
-		}
-		userCount++
-		if firstPrompt == "" {
-			firstPrompt = strings.TrimSpace(message.ContentText)
-		}
-	}
-	if userCount != 1 || firstPrompt == "" {
-		return "", false, nil
-	}
-	return firstPrompt, true, nil
-}
-
-func isDefaultConversationTitle(title string) bool {
-	return strings.TrimSpace(title) == platform.AgentConversationDefaultTitle
-}
-
-func cleanGeneratedTitle(text string) string {
-	text = thinkBlockPattern.ReplaceAllString(text, "")
-	for _, line := range strings.Split(text, "\n") {
-		title := strings.TrimSpace(line)
-		if title == "" {
-			continue
-		}
-		title = strings.TrimSpace(strings.Trim(title, "\"'`*_# \t\r\n"))
-		title = strings.TrimRight(title, ".!?:;")
-		title = strings.TrimSpace(title)
-		if title == "" {
-			continue
-		}
-		runes := []rune(title)
-		if len(runes) > maxGeneratedTitleRunes {
-			title = strings.TrimSpace(string(runes[:maxGeneratedTitleRunes]))
-		}
-		return title
-	}
-	return ""
-}
-
-func fallbackConversationTitle(prompt string) string {
-	title := cleanGeneratedTitle(prompt)
-	if title == "" {
-		return ""
-	}
-	runes := []rune(title)
-	runes[0] = []rune(strings.ToUpper(string(runes[0])))[0]
-	return string(runes)
 }
 
 type PromptInput struct {
@@ -837,42 +703,4 @@ func platformRole(role agent.Role) string {
 
 func systemPrompt() string {
 	return `You are LibreDash's read-only BI assistant. Answer using only the provided tools and conversation context. You can help users understand dashboards, semantic models, metric views, filters, visuals, and table snapshots they are allowed to access. Do not invent dashboard IDs, metric names, or data values. You cannot write data, deploy changes, edit permissions, run raw SQL, access files, or call external services.`
-}
-
-func titleSystemPrompt() string {
-	return `You are a conversation title generator. Output only one title.
-
-<task>
-Generate a brief title that helps the user find this chat later.
-
-Follow all rules in <rules>.
-Use the <examples> so you know what a good title looks like.
-Your output must be:
-- A single line
-- 50 characters or fewer
-- No explanations
-</task>
-
-<rules>
-- Use the same language as the user message you are summarizing
-- Title must be grammatically correct and read naturally
-- Never include tool names, tool calls, model names, or agent internals
-- Focus on the main BI topic or question the user needs to retrieve
-- Preserve exact dashboard names, metric names, model names, IDs, HTTP codes, and numbers
-- Do not answer the user's question
-- Do not use markdown, quotes, or trailing punctuation
-- Do not say you cannot generate a title or complain about the input
-- Always output something meaningful, even if the input is minimal
-- If the user message is short or conversational, create a useful neutral title such as Greeting, Quick check-in, or Intro message
-</rules>
-
-<examples>
-"what dashboards do we have available?" -> Available dashboards
-"show me revenue by month" -> Monthly revenue
-"describe executive-sales" -> Executive Sales dashboard
-"what metrics are in Orders Metrics?" -> Orders Metrics overview
-"why is delivery time so high?" -> Delivery time investigation
-"list all metric views" -> Metric views
-"compare order count and freight value" -> Orders and freight comparison
-</examples>`
 }
